QueryじゃなくてFilterクラスにしたり、名前をわかりやすくした
