+++ src/filter.ts
@@ -1,48 +1,124 @@
 import { Item } from "../utils";
-import { Condition, Query, initial } from "./query";
-
-export type Action = (cond: Condition) => Condition;
-
-export function Reducer(cond: Condition, action: Action): Condition {
-  return action(cond);
-}
-
-export function SetText(s: string): Action {
-  return (cond: Condition): Condition => ({ ...cond, text: s });
-}
-
-export function SetJob(bits: number): Action {
-  return (cond: Condition): Condition => ({
-    ...cond,
-    job_flags: cond.job_flags ^ (1 << bits) /** 指定ジョブのbitを反転させる */
-  });
-}
-
-export function SetSlot(bits: number): Action {
-  return (cond: Condition): Condition => ({
-    ...cond,
-    slot_flags:
-      cond.slot_flags ^ (1 << bits) /** 指定スロットのbitを反転させる */
-  });
-}
-
-export function SetSkill(bits: number): Action {
-  return (cond: Condition): Condition => ({
-    ...cond,
-    skill: cond.skill ^ (1 << bits)
-  });
-}
-
-export function SetMinLevel(level: number): Action {
-  return (cond: Condition): Condition => ({
-    ...cond,
-    minLevel: level
-  });
-}
-
-export const Reset: Action = () => initial();
+import { Condition } from "./condition";
 
 export function Apply(cond: Condition, item: Item[]): Item[] {
-  const q = new Query(cond);
-  return item.filter(x => q.accept(x));
+  const f = new FilterSet(cond);
+  return item.filter(x => f.accept(x));
+}
+
+interface Predicate {
+  accept(item: Item): boolean;
+}
+
+class FilterSet implements Predicate {
+  private ps: Predicate[];
+
+  constructor(cond: Condition) {
+    this.ps = cond.text.split(/\s/).map(x => textFilter(x)) ?? [];
+    if (cond.job_flags) {
+      this.ps.push(jobFilter(cond.job_flags));
+    }
+    if (cond.slot_flags) {
+      this.ps.push(slotFilter(cond.slot_flags));
+    }
+    if (cond.skill) {
+      this.ps.push(skillFilter(cond.skill));
+    }
+    if (cond.minLevel) {
+      this.ps.push(minLevelFilter(cond.minLevel));
+    }
+  }
+
+  accept(item: Item): boolean {
+    return this.ps.every((p: Predicate) => p.accept(item));
+  }
+}
+
+function textFilter(word: string): Predicate {
+  const gte = word.match(/^(.*)>=([-+]?\d+)$/);
+  if (gte) {
+    const ext = propValue(gte[1]);
+    const threshold = Number(gte[2]);
+    return new Filter(ext, x => x >= threshold);
+  }
+  const gt = word.match(/^(.*)>([-+]?\d+)$/);
+  if (gt) {
+    const ext = propValue(gt[1]);
+    const threshold = Number(gt[2]);
+    return new Filter(ext, x => x > threshold);
+  }
+  const lte = word.match(/^(.*)<=([-+]?\d+)$/);
+  if (lte) {
+    const ext = propValue(lte[1]);
+    const threshold = Number(lte[2]);
+    return new Filter(ext, x => x <= threshold);
+  }
+  const lt = word.match(/^(.*)<([-+]?\d+)$/);
+  if (lt) {
+    const ext = propValue(lt[1]);
+    const threshold = Number(lt[2]);
+    return new Filter(ext, x => x < threshold);
+  }
+  const eq = word.match(/^(.*)=([-+]?\d+)$/);
+  if (eq) {
+    const ext = propValue(eq[1]);
+    const value = Number(eq[2]);
+    return new Filter(ext, x => x === value);
+  }
+  return new Filter(wholeText, x => x.includes(word));
+}
+
+type extractor<T> = (item: Item) => T | null;
+
+class Filter<T> implements Predicate {
+  extract: extractor<T>;
+  f: (x: T) => boolean;
+  constructor(ext: extractor<T>, f: (x: T) => boolean) {
+    this.extract = ext;
+    this.f = f;
+  }
+
+  accept(item: Item): boolean {
+    const x = this.extract(item);
+    return x ? this.f(x) : false;
+  }
+}
+
+export function propValue(prop: string): extractor<number> {
+  const re = new RegExp(`(^|\\s)${prop}([-+]\\d+)`);
+  return (item: Item) => {
+    const x = item.description.match(re);
+    return x ? Number(x[2]) : null;
+  };
+}
+
+const wholeText: extractor<string> = (item: Item) =>
+  [item.name, item.description].join();
+
+function jobFilter(job_flags: number): Filter<number> {
+  return new Filter(
+    item => item._jobs,
+    that => (that & job_flags) > 0
+  );
+}
+
+function slotFilter(slot_flags: number): Filter<number> {
+  return new Filter(
+    item => item._slots,
+    that => (that & slot_flags) > 0
+  );
+}
+
+function skillFilter(skill: number): Filter<number> {
+  return new Filter(
+    item => item.skill,
+    id => ((1 << id) & skill) > 0
+  );
+}
+
+function minLevelFilter(level: number): Filter<number> {
+  return new Filter(
+    item => item.item_level ?? item.level,
+    that => that >= level
+  );
 }
